+++ site/pub/guide/27.js
@@ -121,7 +121,6 @@ scene("gameover", (score) => {
 		origin("center"),
 	]);
 
-
 	keyPress("space", () => {
 		go("game");
 	});