+++ site/pub/guide/28.js
@@ -124,7 +124,6 @@ scene("gameover", (score) => {
 		origin("center"),
 	]);
 
-
 	keyPress("space", () => {
 		go("game");
 	});